refactor: 将 Props 放置到组件文件内，移除组件目录中的 `types.ts`
