--- src/button/types.ts
@@ -1,18 +1,0 @@
-import type { Component, Snippet } from "svelte";
-import type { ClassValue, HTMLButtonAttributes } from "svelte/elements";
-import type { ColorType, IconPosition, SizeType } from "../types";
-
-export interface BaseButtonProps extends HTMLButtonAttributes {
-  class?: ClassValue;
-  onclick?: () => void;
-  loading?: boolean;
-  disabled?: boolean;
-  children: Snippet;
-}
-
-export interface ButtonProps extends BaseButtonProps {
-  size?: SizeType;
-  color?: ColorType;
-  icon?: Component;
-  iconPosition?: IconPosition;
-}
