+++ src/placeholder/Placeholder.svelte
@@ -8,10 +8,8 @@
     color?: ColorType;
   }
 
-  const {
-    class: className = "",
-    color = "primary",
-  }: PlaceholderProps = $props();
+  const { class: className = "", color = "primary" }: PlaceholderProps =
+    $props();
 </script>
 
 <div
