--- src/tabs/types.ts
@@ -1,8 +1,0 @@
-import type { Component } from "svelte";
-
-export interface TabItem<T> {
-  name: string;
-  value: T;
-  icon?: Component;
-  disabled?: boolean;
-}
